Add attempt at invoking less manually
After some time playing about with both displaying single lined tables
properly in the pager and with invoking less manually to exit if less
than one screen of material is displayed I'm having to admit defeat.

For posterity the issue with the single lined tables still eludes me as
it seems to display unicode chars fine
With invoking less manually the issue is based in the fact that it takes
passing the table to the command which exceeds the max length for bash
argument. This may be fixable (see
https://www.linuxjournal.com/article/6060?page=0,0) but I don't feel
qualifed at present