Stop prompt-mode user modify requiring an ssh key
Previously if in simple mode & modifying a user via the prompts system,
if the user did not have an ssh key set & a new value wasn't
passed the system would try to pass "None" (as a string) as its value.
This would neccessarily produce an "ipa: ERROR: invalid 'sshpubkey':
invalid SSH public key" error as 'None' is not a valid ssh_key.

This system has now been changed to only pass the value if it's not the
default